Create S3_To_DynamoDB.py
This code is used to transfer the contents of a .txt file in AWS S3 Bucket to A corresponding dynamoDB table in AWS using a Python code. For reference, we have taken the table name to be 'employee' with 3 attributes - "Id", "name", "salary". The code can be generalized by making changes to it as per the requirement and is hence a great reference.