functionpointer
Have a set of functions defined multi, addit, modul and divi which does the corresponding operations of Multiplication, Addition, Remainder calculation and Division for two integers respectively. Give an option to the user for selecting the function to execute and use function pointers for execution.
Eg. Enter any of the following functions to execute

multi, addit, modul and divi
Input:

multi

enter the numbers

2 

4

Answer: 6